+++ packages/ui/src/proposals/modals/AddNewProposal/components/SpecificParameters/WorkingGroupLeadOpening/CreateWorkingGroupLeadOpening/DurationAndProcess.tsx
@@ -1,11 +1,9 @@
 import React from 'react'
 
-import { CKEditor } from '@/common/components/CKEditor'
-import { InputComponent } from '@/common/components/forms'
 import { Row } from '@/common/components/Modal'
-import { OpeningDuration } from '@/common/components/OpeningDuration/OpeningDuration'
 import { RowGapBlock } from '@/common/components/page/PageContent'
-import { TextBig, TextMedium } from '@/common/components/typography'
+import { TextMedium } from '@/common/components/typography'
+import { DurationAndProcess as DurationAndProcess_ } from '@/working-groups/components/CreateOpening/DurationAndProcess'
 
 export const DurationAndProcess = () => {
   return (
@@ -18,22 +16,7 @@ export const DurationAndProcess = () => {
       </Row>
 
       <Row>
-        <RowGapBlock gap={20}>
-          <TextBig bold value>
-            Opening Duration
-          </TextBig>
-          <OpeningDuration label="Expected length of the application period" />
-
-          <InputComponent
-            label="Application process"
-            required
-            inputSize="auto"
-            id="field-details"
-            name="durationAndProcess.details"
-          >
-            <CKEditor id="field-details" name="durationAndProcess.details" />
-          </InputComponent>
-        </RowGapBlock>
+        <DurationAndProcess_ />
       </Row>
     </RowGapBlock>
   )
