+++ packages/ui/src/proposals/modals/AddNewProposal/components/SpecificParameters/WorkingGroupLeadOpening/CreateWorkingGroupLeadOpening/WorkingGroupAndDescription.tsx
@@ -1,13 +1,9 @@
-import React, { useEffect } from 'react'
-import { useFormContext } from 'react-hook-form'
+import React from 'react'
 
-import { CKEditor } from '@/common/components/CKEditor'
-import { InputComponent, InputText, InputTextarea } from '@/common/components/forms'
 import { Row } from '@/common/components/Modal'
 import { RowGapBlock } from '@/common/components/page/PageContent'
 import { TextMedium } from '@/common/components/typography'
-import { SelectWorkingGroup } from '@/working-groups/components/SelectWorkingGroup'
-import { GroupIdToGroupParam } from '@/working-groups/constants'
+import { WorkingGroupAndDescription as WorkingGroupAndDescription_ } from '@/working-groups/components/CreateOpening/WorkingGroupAndDescription'
 import { GroupIdName } from '@/working-groups/types'
 
 interface Props {
@@ -15,18 +11,6 @@ interface Props {
 }
 
 export const WorkingGroupAndDescription = ({ groupId }: Props) => {
-  const { watch, setValue } = useFormContext()
-  const MappedGroupedParams = {
-    ...GroupIdToGroupParam,
-    operationsWorkingGroupAlpha: 'Builders',
-    operationsWorkingGroupBeta: 'HR',
-    operationsWorkingGroupGamma: 'Marketing',
-  }
-
-  useEffect(() => {
-    setValue('workingGroupAndDescription.groupId', groupId, { shouldValidate: true })
-    setValue('groupId', groupId)
-  }, [groupId])
 
   return (
     <RowGapBlock gap={24}>
@@ -37,55 +21,7 @@ export const WorkingGroupAndDescription = ({ groupId }: Props) => {
         </RowGapBlock>
       </Row>
       <Row>
-        <RowGapBlock gap={20}>
-          {groupId ? (
-            <div>Group: {MappedGroupedParams[groupId]}</div>
-          ) : (
-            <InputComponent
-              id="working-group-select"
-              label="Working Group"
-              required
-              inputSize="l"
-              tooltipText="Please select an identifier for working group"
-            >
-              <SelectWorkingGroup
-                id="working-group-select"
-                selectedGroupId={watch('workingGroupAndDescription.groupId')}
-                onChange={(selected) => {
-                  setValue('workingGroupAndDescription.groupId', selected.id, { shouldValidate: true })
-                  setValue('groupId', selected.id)
-                }}
-              />
-            </InputComponent>
-          )}
-          <InputComponent
-            id="opening-title"
-            label="Opening title"
-            required
-            inputSize="m"
-            name="workingGroupAndDescription.title"
-          >
-            <InputText id="opening-title" name="workingGroupAndDescription.title" placeholder="Opening title" />
-          </InputComponent>
-          <InputComponent
-            id="short-description"
-            label="Short description"
-            name="workingGroupAndDescription.shortDescription"
-            required
-            inputSize="l"
-          >
-            <InputTextarea id="short-description" name="workingGroupAndDescription.shortDescription" />
-          </InputComponent>
-          <InputComponent
-            label="Description"
-            required
-            inputSize="auto"
-            id="field-description"
-            name="workingGroupAndDescription.description"
-          >
-            <CKEditor id="field-description" name="workingGroupAndDescription.description" />
-          </InputComponent>
-        </RowGapBlock>
+        <WorkingGroupAndDescription_ groupId={groupId} />
       </Row>
     </RowGapBlock>
   )
