+++ packages/ui/src/working-groups/components/CreateOpening/DurationAndProcess.tsx
